backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Google Map restaurant finder app"
authors = ["Adam Hoffstein <adam@C0377-Macbook-Pro.local>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
psycopg2-binary = "^2.9.3"
googlemaps = "^4.6.0"
SQLAlchemy = "^1.4.32"
alembic = "^1.7.7"
python-dotenv = "^0.20.0"
pandas = "^1.4.1"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
pytest = "^7.1.1"
python-multipart = "^0.0.5"
pyflakes = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
